chore: rename phoenix-flow integration to baton
Tracks the upstream rename phoenixvc/phoenix-flow → phoenixvc/baton
(2026-04-26). Mechanical rename across the MCP integration layer.

Renames:
- lib/integrations/phoenix-flow.ts → lib/integrations/baton.ts
- PhoenixFlowConfig → BatonConfig
- PhoenixFlowFeatureFlag → BatonFeatureFlag
- PhoenixFlowUnavailableError → BatonUnavailableError
- isPhoenixFlowEnabled() → isBatonEnabled()
- featureFlags.phoenixFlow → featureFlags.baton
- PHOENIX_FLOW_MCP_URL → BATON_MCP_URL
- PHOENIX_FLOW_MCP_SECRET → BATON_MCP_SECRET

Rebased onto current main (after #123 prettier landed) — branch was
hard-reset to origin/main and the rename was re-applied cleanly so
the diff is purely the rename, with no formatting churn.

Breaking for anyone with a saved data/feature-flags.json or .env.local —
re-key from phoenixFlow / PHOENIX_FLOW_* to baton / BATON_*.
