+++ .env.example
@@ -122,12 +122,13 @@ JWT_SECRET=your-secure-jwt-secret-key-here
 # IDENTITY_API_KEY=your-identity-api-key
 
 # ============================================
-# PHOENIX-FLOW TASK MANAGEMENT (Optional)
+# BATON TASK MANAGEMENT (Optional)
 # ============================================
-# Phoenix-Flow MCP server for task management with org context
-# Requires the phoenixFlow feature flag to be enabled
-# PHOENIX_FLOW_MCP_URL=https://your-phoenix-flow.railway.app/mcp
-# PHOENIX_FLOW_MCP_SECRET=your-mcp-secret
+# Baton MCP server for task management with org context
+# (Baton is the renamed phoenix-flow service)
+# Requires the `baton` feature flag to be enabled
+# BATON_MCP_URL=https://your-baton.railway.app/mcp
+# BATON_MCP_SECRET=your-mcp-secret
 # MCP_ORG_URL=https://your-mcp-org.railway.app
 # MCP_ORG_SECRET=your-org-secret
